pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinelops"
version = "1.0.0"
description = "SentinelOps — Multi-step agentic RL environment for AI security surveillance"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Team Adaptrix"},
]
keywords = ["openenv", "reinforcement-learning", "surveillance", "security", "agentic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "openai>=1.12.0",
    "requests>=2.31.0",
    "Pillow>=10.2.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["server*", "tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
